LayoutTests/imported/w3c/web-platform-tests/html/syntax/speculative-parsing/generated/document-write/link-rel-stylesheet-referrerpolicy-no-referrer.tentative.sub-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_equals: expected the same number of headers between speculative and non-speculative expected 2 but got 3

    <!-- speculative case in document.write -->
    <link rel=stylesheet href="/html/syntax/speculative-parsing/resources/stash.py?action=put&amp;uuid=fec864e8-1a9a-427f-9022-45cd300bd161&amp;encodingcheck=&Gbreve;" referrerpolicy=no-referrer>



FAIL Speculative parsing, document.write(): link-rel-stylesheet-referrerpolicy-no-referrer Unhandled rejection: assert_equals: expected the same number of headers between speculative and non-speculative expected 2 but got 3
